requirements.txt
pyyaml==6.0
pydantic==1.10.7
pysnmp==4.4.12